cleaned up and defined metadata types
