--- js/sdks/packages/secure-frame-iframe/src/types.ts
@@ -1,7 +1,0 @@
-export interface MetaData {
-  fileinfo: {
-    filename: string;
-    type?: string;
-    lastModified?: number;
-  };
-}
